app/requirements.txt
streamlit 1.26.0
tensorflow 2.10.0
Pillow 9.5.0
mtcnn 0.1.1
requests 2.31.0
opencv-python 4.8.0.76